Use OpenSceneGraph for rendering instead of Ogre
Some big changes here.

A RenderManager for MyGUI, using OSG, is added, and along with a fixed up
PhysFSDataManager and the default LogManager replaces the MyGUI's Ogre
Platform.

An SDL2-based GraphicsWindow class is added to allow OSG to render to the SDL
window. GraphicsWindowSDL2 is slave based, meaning it only works with a premade
SDL_Window (which must be correctly set up for OpenGL rendering) and does not
check SDL events. The app is responsible for handling events and passing them
on to OSG as needed (resize events in particular).

Terrain rendering is broken. It needs to be rewritten for OSG.

FPS and stats display is broken. Will be fixed when I figure out how to
properly calculate the framerate from OSG.