Use a deferred shading pipeline to render the scene
As usual, many changes here. Biggest being that it's now fully a shader-based
pipeline, not using the fixed-function pipeline (though it does make use of
OSG's matrix and vertex attribute aliasing features to get the state into the
shaders).

The MyGUI RenderManager has been changed to use shaders to accomdate.

The terrain system has reimplemented the shader path, with the fragment shader
being generated dynamically. Newer OSG versions should have a way of using
mark-up syntax in shader files to have the base code in an external file, but
for now it's generated in code.

Unfortunately it now requires hardware that supports 5 simultaneous render
targets. At least 4 are definitely needed for self-illuminated geometry to
work, with the 5th (specular light buffer) only needed so it can be cleared
with the diffuse light buffer. A separate clear-only path could theoretically
work around the need for 5.